Add initial code for pulse counting
Brief intro: my stationary bicycle *had* a measurement monitor with
speed, traveled kms, etc. However, it broke down. As my stationary bike
is quite old, I was not able to find a replacement (a good thing). I
noticed that the cable going from the rotor to the monitoring device has
a 3.5mm jack ending. I put it into the microphone slot on my computer
and started pedalling. I noticed that after each finished circle (round,
or candence or how it's called), it made a beeping or *pulsing* sound.
Long story short, I made a simple Python program that monitors the
microphone and tries to detect the *pulses*. If it detect the pulse, it
simply prints out "pulse". When you know the distance you traveled in
each *pulse*, you can calculate almost anything (speed, distance, etc).
I'm going to create a simple React-based UI for stationary bicycles
based on this discovery.

In this commit, I'm adding a simple python program to detect the pulses
form the stationary bicycle.